renaming files to snake case, moving files
