+++ src/domains/audio-player/audio-progress/audio-progress.component.js
