+++ src/domains/audio-player/audio-progress/audio-progress.utils.js
