+++ src/domains/audio-player/audio-progress/index.js
@@ -1,0 +1,1 @@
+export { default } from './audio-progress.container';
