+++ src/domains/audio-player/audio-progress/styles/progress.scss
