+++ src/domains/audio-player/buttons/forward/forward-button.test.js
@@ -2,14 +2,14 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import { render } from '@testing-library/react';
 
-import ForwardButton from 'domains/audio/playerButtons/forward';
-
 import store from 'store';
+import ForwardButton from './forward-button';
 
 test('It should render', () => {
+  const mockFn = jest.fn();
   render(
     <Provider store={store}>
-      <ForwardButton />
+      <ForwardButton onClick={mockFn} />
     </Provider>,
   );
 });
