+++ src/domains/audio-player/buttons/forward/index.js
@@ -1,0 +1,1 @@
+export { default } from './forward-button';
