+++ src/domains/audio-player/buttons/play/index.js
@@ -1,0 +1,1 @@
+export { default } from './play-button';
