+++ src/domains/audio-player/buttons/repeat/index.js
@@ -1,0 +1,3 @@
+import RepeatButton from './repeat-button';
+
+export default RepeatButton;
