+++ src/domains/audio-player/buttons/rewind/index.js
@@ -1,0 +1,1 @@
+export { default } from './rewind-button';
