+++ src/domains/audio-player/buttons/rewind/rewind-button.test.js
@@ -2,14 +2,15 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import { render } from '@testing-library/react';
 
-import RewindButton from 'domains/audio/playerButtons/rewind';
-
 import store from 'store';
+import RewindButton from './rewind-button';
 
 test('It should render', () => {
+  const mockFn = jest.fn();
+
   render(
     <Provider store={store}>
-      <RewindButton />
+      <RewindButton onClick={mockFn} />
     </Provider>,
   );
 });
