+++ src/domains/audio-player/buttons/shuffle/index.js
@@ -1,0 +1,3 @@
+import ShuffleButton from './shuffle-button';
+
+export default ShuffleButton;
