+++ src/domains/audio-player/index.js
@@ -1,0 +1,3 @@
+import AudioPlayerContainer from './audio-player.container';
+
+export default AudioPlayerContainer;
