+++ src/domains/audio-player/style/player.scss
