--- src/domains/audio/audioProgress/index.js
@@ -1,1 +1,0 @@
-export { default } from './audioProgress.container';
