--- src/domains/audio/index.js
@@ -1,3 +1,0 @@
-import AudioPlayerContainer from './audioPlayer.container';
-
-export default AudioPlayerContainer;
