--- src/domains/audio/playerButtons/forward/index.js
@@ -1,1 +1,0 @@
-export { default } from './forwardButton';
