--- src/domains/audio/playerButtons/play/index.js
@@ -1,1 +1,0 @@
-export { default } from './playButton';
