--- src/domains/audio/playerButtons/repeat/index.js
@@ -1,3 +1,0 @@
-import RepeatButton from './repeatButton';
-
-export default RepeatButton;
