--- src/domains/audio/playerButtons/rewind/index.js
@@ -1,1 +1,0 @@
-export { default } from './rewindButton';
