--- src/domains/audio/playerButtons/shuffle/index.js
@@ -1,3 +1,0 @@
-import ShuffleButton from './shuffleButton';
-
-export default ShuffleButton;
